updated front end to line up with changes in database on backend
