+++ controllers/gameController.js
@@ -6,10 +6,28 @@ async function get(req, res) {
   try {
     const games = await db.Game.find({ name: req.params.name });
     if(games.length === 0) {
-      const results = await client.games({search: req.params.name, filters: {'popularity-gt': '5'}, fields: ['name', 'summary', 'genres', 'developers', 'publishers', 'release_dates', 'game_modes', 'rating', 'esrb', 'cover'], expand: ['developers', 'publishers', 'game_modes', 'genres']});  
-      res.status(200).json(results.body);
+      function precisionRound(number, precision) {
+        var factor = Math.pow(10, precision);
+        return Math.round(number * factor) / factor;
+      }
+      const results = await client.games({search: req.params.name, filters: {'popularity-gt': '5'}, fields: ['name', 'summary', 'genres', 'developers', 'publishers', 'release_dates', 'game_modes', 'rating', 'esrb', 'cover'], expand: ['developers', 'publishers', 'game_modes', 'genres']});
+      const newGames = results.body.map(game => {
+        return {
+          name: game.name,
+          summary: game.summary,
+          genre: game.genres.map(genre => genre.name),
+          developer: game.developers.map(developer => ({name: developer.name, url: developer.website})),
+          publisher: game.publishers.map(publisher => ({name: publisher.name, url: publisher.website})),
+          releaseDate: game.release_dates ? game.release_dates[0].human : null, 
+          gameMode: game.game_modes.map(gameMode => gameMode.name),         
+          rating: precisionRound(game.rating, 1),
+          esrb: game.esrb ? game.esrb.rating : null,
+          coverPhoto: `https://images.igdb.com/igdb/image/upload/t_720p/${game.cover.cloudinary_id}.jpg`,
+        }
+      });  
+      res.status(200).json({ new: true, games: newGames });
     }
-    res.status(200).json(games);
+    res.status(200).json({ new: false, games});
   } catch(err) {
     console.log(err);
   }
