+++ routes/postRouter.js
@@ -3,7 +3,7 @@ const postController = require('../controllers/postController');
 
 const router = express.Router();
 
-router.route('/api/post').get(postController.get);
+router.route('/api/post/:id').get(postController.get);
 router.route('/api/post').post(postController.post);
 
 module.exports = router;
